refactor: organize files
* remove `gridWithMouse`
* remove `panelContainer` hooks
* remove `gridSelection` hooks
* organize and move hooks directory
* rename hooks
