+++ apps/calendar/src/components/dayGridWeek/otherGridRow.tsx
@@ -4,11 +4,11 @@ import range from 'tui-code-snippet/array/range';
 
 import { GridCells } from '@src/components/dayGridWeek/gridCells';
 import { HorizontalEvent } from '@src/components/events/horizontalEvent';
-import { useHeightController } from '@src/components/hooks/heightController';
 import Template from '@src/components/template';
 import { PANEL_HEIGHT, WEEK_EVENT_MARGIN_TOP } from '@src/constants/style';
 import { cls } from '@src/helpers/css';
 import { EVENT_HEIGHT, isWithinHeight } from '@src/helpers/grid';
+import { useGridRowHeightController } from '@src/hooks/dayGridWeek/gridRowHeightController';
 import EventUIModel from '@src/model/eventUIModel';
 import TZDate from '@src/time/date';
 import { addDate } from '@src/time/datetime';
@@ -53,7 +53,7 @@ export const OtherGridRow: FunctionComponent<Props> = ({
   const columnWidth = timesWidth * timezonesCount;
 
   const { clickedIndex, isClickedCount, onClickExceedCount, onClickCollapseButton } =
-    useHeightController(maxTop, category);
+    useGridRowHeightController(maxTop, category);
 
   const horizontalEvents = events
     .filter(isWithinHeight(height, EVENT_HEIGHT + WEEK_EVENT_MARGIN_TOP))
