+++ apps/calendar/src/components/popup/eventStateSelector.tsx
@@ -4,10 +4,9 @@ import { useState } from 'preact/hooks';
 import { PopupSection } from '@src/components/popup/popupSection';
 import { StateDropdownMenu } from '@src/components/popup/stateDropdownMenu';
 import { cls } from '@src/helpers/css';
+import { useDropdownState } from '@src/hooks/common/dropdownState';
 
 import { EventState } from '@t/events';
-
-import { useDropdownState } from '../hooks/dropdownState';
 
 interface Props {
   eventState: EventState;
